Remove make_matrix, build into ckmeans
