--- src/make_matrix.js
@@ -1,24 +1,0 @@
-'use strict';
-
-/**
- * Create a new column x row matrix.
- *
- * @param {number} columns
- * @param {number} rows
- * @return {Array<Array<number>>} matrix
- * @example
- * makeMatrix(10, 10);
- */
-function makeMatrix(columns, rows) {
-    var matrix = [];
-    for (var i = 0; i < columns; i++) {
-        var column = [];
-        for (var j = 0; j < rows; j++) {
-            column.push(0);
-        }
-        matrix.push(column);
-    }
-    return matrix;
-}
-
-module.exports = makeMatrix;
